changed cocktail to drink, renderLineItems needed second parameter, logging ids for products
