+++ products/index.html
@@ -22,6 +22,10 @@
         <section class="main-section">
             Check out my products!
             <ul id="list"></ul>
+            <div class="dropdown">
+                <button onclick="quantitySelect()" class="dropButton">Dropdown</button>
+                <div id="qtyDropdown"></div>
+            </div>
             <a href="../shopping-cart">Shopping Cart</a>
         </section>
     </main>
