+++ shopping-cart/index.html
@@ -20,7 +20,7 @@
     </header>
     <main>
         <section class="main-section">
-            Welcome to the cart page!
+            Welcome to the shopping cart!
             <table>
                 <thead>
                     <tr>
