+++ shopping-cart/shopping-cart-utils.js
@@ -1,13 +1,5 @@
-import { findById, getFromLocalStorage } from '../utils.js';
+import { findById } from '../utils.js';
 import { cocktails } from '../data/cocktails.js';
-
-const orderButton = document.querySelector('#order');
-
-const cart = getFromLocalStorage('CART') || [];
-
-if (cart.length === 0) {
-    orderButton.disabled = true;
-}
 
 export function calcOrderTotal(cartArray) {
     let total = 0;
@@ -24,11 +16,3 @@ export function calcOrderTotal(cartArray) {
 
     return total;
 }
-
-orderButton.addEventListener('click', () => {
-    const stringyCart = JSON.stringify(cart, true, 2);
-    alert(stringyCart);
-
-    localStorage.clear();
-    window.location.href = '/';
-});
