+++ shopping-cart/shopping-cart.js
@@ -1,17 +1,16 @@
-import { cart } from '../data/cart.js';
 import { calcOrderTotal } from '../shopping-cart/shopping-cart-utils.js';
 import { renderLineItems } from '../shopping-cart/render-line-items.js';
-
+import { getFromLocalStorage } from '../utils.js';
+import { cocktails } from '../data/cocktails.js';
 
 const table = document.querySelector('tbody');
+const cart = getFromLocalStorage('CART') || [];
 
 for (let i = 0; i < cart.length; i++) {
-    const cocktail = cart[i];
+    const drink = cart[i];
 
-    if (cocktail.quantity >= 0) {
-        const tr = renderLineItems(cocktail);
-        table.appendChild(tr);
-    }
+    const tr = renderLineItems(drink, cocktails);
+    table.appendChild(tr);
 }
 
 const total = calcOrderTotal(cart);
@@ -19,3 +18,17 @@ const total = calcOrderTotal(cart);
 const totalLine = document.querySelector('.total');
 
 totalLine.textContent = `Total: $${total}`;
+
+const orderButton = document.querySelector('#order');
+
+if (cart.length === 0) {
+    orderButton.disabled = true;
+}
+
+orderButton.addEventListener('click', () => {
+    const stringyCart = JSON.stringify(cart, true, 2);
+    alert(stringyCart);
+
+    localStorage.clear();
+    window.location.href = '/';
+});
